Fix DropDownView
1) The classes related to the displayed interfacing have been moved to utills/ytpromo.py. Tests were performed, no errors were detected
